Clean up quicksort CLI

+++ src/quicksort.rs
@@ -2,7 +2,7 @@ use crate::short_canonical_game::{Game, GameBackend, Moves};
 use std::fmt::Display;
 
 #[derive(Debug, PartialEq, Eq)]
-pub struct Quicksort(pub Vec<usize>);
+pub struct Quicksort(pub Vec<u32>);
 
 impl Display for Quicksort {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
